@@ -8,32 +8,31 @@
     "build:prod": "ng build --configuration production --aot && cp .htaccess ../../../public && cp index.php ../../../public",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "postinstall": "ngcc"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.11",
-    "@angular/common": "^13.3.11",
-    "@angular/compiler": "^13.3.11",
-    "@angular/core": "^13.3.11",
-    "@angular/forms": "^13.3.11",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
-    "@datorama/akita": "^6.1.3",
+    "@angular/animations": "^14.2.0",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
     "@types/lodash": "^4.14.168",
     "lodash": "^4.17.21",
-    "ng2-order-pipe": "^0.1.5",
-    "ng2-search-filter": "^0.5.1",
-    "ngx-pagination": "^5.1.1",
-    "rxjs": "^6.6.7",
-    "tslib": "^2.0.0",
+    "ngx-bootstrap": "^8.0.0",
+    "ngx-scrollbar": "^9.0.2",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.9",
-    "@angular/cli": "^13.3.9",
-    "@angular/compiler-cli": "^13.3.11",
+    "@angular-devkit/build-angular": "^14.2.1",
+    "@angular/cli": "^14.2.1",
+    "@angular/compiler-cli": "^14.2.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.20.10",
     "codelyzer": "^6.0.0",
@@ -48,6 +47,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.1.6"
+    "typescript": "~4.6.2"
   }
 }
