@@ -4,6 +4,5 @@
 
 export const environment = {
   production: false,
-  baseUrl: "http://localhost:8000/api/",
-  showErrorTrace:true,
+  baseUrl: "http://localhost:8000/api/"
 };
